Separate macOS bootstrap state model
